- status removed from badly written database cache; - code cleaning
git-svn-id: http://svn.miranda-ng.org/main/trunk@17162 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/TabSRMM/src/container.cpp
@@ -1876,7 +1876,7 @@ void TSAPI CloseOtherTabs(HWND hwndTab, TWindowData &dat)
 
 int TSAPI CutContactName(const wchar_t *oldname, wchar_t *newname, size_t size)
 {
-	if (mir_wstrlen(oldname) <= PluginConfig.m_iTabNameLimit)
+	if ((int)mir_wstrlen(oldname) <= PluginConfig.m_iTabNameLimit)
 		wcsncpy_s(newname, size, oldname, _TRUNCATE);
 	else {
 		wchar_t fmt[30];
